+++ pysrc/juliacall/__init__.py
@@ -1,7 +1,7 @@
 # This module gets modified by PythonCall when it is loaded, e.g. to include Core, Base
 # and Main modules.
 
-__version__ = '0.9.3'
+__version__ = '0.9.4'
 
 _newmodule = None
 
@@ -20,6 +20,13 @@ def convert(T, x):
     if _convert is None:
         _convert = PythonCall.seval("pyjlcallback((T,x)->pyjl(pyconvert(pyjlvalue(T)::Type,x)))")
     return _convert(T, x)
+
+def interactive(enable=True):
+    "Allow the Julia event loop to run in the background of the Python REPL."
+    if enable:
+        PythonCall._set_python_input_hook()
+    else:
+        PythonCall._unset_python_input_hook()
 
 class JuliaError(Exception):
     "An error arising in Julia code."
@@ -122,6 +129,7 @@ def init():
     CONFIG['opt_sysimage'] = sysimg = path_option('sysimage', check_exists=True)[0]
     CONFIG['opt_threads'] = int_option('threads', accept_auto=True)[0]
     CONFIG['opt_warn_overwrite'] = choice('warn_overwrite', ['yes', 'no'])[0]
+    CONFIG['opt_handle_signals'] = 'no'
 
     # Stop if we already initialised
     if CONFIG['inited']:
@@ -178,6 +186,7 @@ def init():
             return 'raw"' + x.replace('"', '\\"').replace('\\', '\\\\') + '"'
         script = '''
         try
+            Base.require(Main, :CompilerSupportLibraries_jll)
             import Pkg
             ENV["JULIA_PYTHONCALL_LIBPTR"] = {}
             ENV["JULIA_PYTHONCALL_EXE"] = {}