+++ docs/src/releasenotes.md
@@ -1,6 +1,15 @@
 # Release Notes
 
 ## Unreleased
+* Experimental new function `juliacall.interactive()` allows the Julia async event loop to
+  run in the background of the Python REPL.
+* Experimental new IPython extension `juliacall.ipython` providing the `%jl` and `%%jl`
+  magics for executing Julia code.
+* Experimental new module `juliacall.importer` allowing you to write Python modules in
+  Julia.
+* Bug fixes.
+
+## 0.9.4 (2022-07-26)
 * Bug fixes.
 
 ## 0.9.3 (2022-07-02)